feat: PostService
관리자 권한으로 수정 삭제 기능 추가

+++ src/main/java/com/example/report/service/PostService.java
@@ -83,11 +83,13 @@ public class PostService {
                 () -> new IllegalArgumentException("사용자가 존재하지 않습니다.")
             );
 
-            Post post = postRepository.findByIdAndId(id, user.getId()).orElseThrow(
+            Post post = postRepository.findById(id).orElseThrow(
                     () -> new IllegalArgumentException("게시물이 존재하지 않습니다.")
             );
-            if (user.getRole().equals(UserRoleEnum.ADMIN) || post.getUsername().equals(user.getUsername()))
-            post.update(requestDto, user);
+
+            if (user.getRole().equals(UserRoleEnum.ADMIN) || post.getUser().getUsername().equals(user.getUsername())){
+                post.update(requestDto, user);
+            }else return null;
 
             return ResponseDto.success("수정 완료");
         } else return ResponseDto.fail(400, "Token Error");
@@ -121,11 +123,13 @@ public class PostService {
                     () -> new IllegalArgumentException("사용자가 존재하지 않습니다.")
             );
 
-            Post post = postRepository.findByIdAndId(id, user.getId()).orElseThrow(
+            Post post = postRepository.findById(id).orElseThrow(
                     () -> new IllegalArgumentException("게시물이 존재하지 않습니다.")
             );
 
-            postRepository.delete(post);
+            if (user.getRole().equals(UserRoleEnum.ADMIN) || post.getUser().getUsername().equals(user.getUsername())){
+                postRepository.delete(post);
+            }else return null;
 
             return ResponseDto.success("삭제 완료");
         } else return ResponseDto.fail(400, "Token Error");
